+++ apps/console/src/layouts/error-layout.tsx
@@ -17,13 +17,13 @@
  */
 
 import { ProtectedRoute } from "@wso2is/admin.core.v1/components";
-import { getErrorLayoutRoutes } from "@wso2is/admin.core.v1/configs";
 import { AppConstants } from "@wso2is/admin.core.v1/constants";
 import { RouteInterface } from "@wso2is/core/models";
 import { ContentLoader, ErrorLayout as ErrorLayoutSkeleton } from "@wso2is/react-components";
 import React, { FunctionComponent, PropsWithChildren, ReactElement, Suspense, useEffect, useState } from "react";
 import { StaticContext } from "react-router";
 import { Redirect, Route, RouteComponentProps, Switch } from "react-router-dom";
+import { getErrorLayoutRoutes } from "../configs/routes";
 
 /**
  * Error layout Prop types.
@@ -43,7 +43,7 @@ export interface ErrorLayoutPropsInterface {
  *
  * @returns a React.ReactElement.
  */
-export const ErrorLayout: FunctionComponent<PropsWithChildren<ErrorLayoutPropsInterface>> = (
+const ErrorLayout: FunctionComponent<PropsWithChildren<ErrorLayoutPropsInterface>> = (
     props: PropsWithChildren<ErrorLayoutPropsInterface>
 ): ReactElement => {
 
@@ -99,3 +99,5 @@ export const ErrorLayout: FunctionComponent<PropsWithChildren<ErrorLayoutPropsIn
 ErrorLayout.defaultProps = {
     fluid: true
 };
+
+export default ErrorLayout;