+++ features/admin.api-resources.v2/pages/api-resources-internal-list.tsx
@@ -165,11 +165,14 @@ const APIResourcesPage: FunctionComponent<APIResourcesPageInterface> = (
             ? `type eq ${ APIResourceCategories.TENANT }`
             : `type eq ${ APIResourceCategories.ORGANIZATION }`;
 
-        if (searchQuery) {
-            setFilter(`${ searchQuery } and ${ typeFilter }`);
-        } else {
-            setFilter(typeFilter);
-        }
+        const apiResourceFilter: string = searchQuery ? `${ searchQuery } and ${ typeFilter }` : typeFilter;
+
+        // Setting `before` and `after` cursors to undefined before initiating a fresh
+        // search query, otherwise text search doesn't work when it is initiated from a
+        // page greater than one.
+        setBefore(undefined);
+        setAfter(undefined);
+        setFilter(apiResourceFilter);
     }, [ searchQuery ]);
 
     /**