+++ features/admin.server-configurations.v1/constants/server-configurations-constants.ts
@@ -432,4 +432,5 @@ export class ServerConfigurationsConstants {
 	 */
 	public static readonly ORGANIZATION_SETTINGS_CATEGORY_ID: string = "organization-settings";
 	public static readonly EMAIL_DOMAIN_DISCOVERY: string = "ZW1haWwtZG9tYWluLWRpc2NvdmVyeQ==";
+	public static readonly IMPERSONATION: string = "impersonation";
 }