+++ modules/i18n/src/translations/en-US/portals/applications.ts
@@ -1145,21 +1145,20 @@ export const applications: ApplicationsNS = {
                     },
                     fields: {
                         enableFIDOTrustedApps: {
-                            hint: "Select to trust the app for user login with passkey. Provide the details of the application under ",
-                            backLink: "Platform Settings.",
+                            hint: "Select to trust the app for user login with passkey. Provide the details of the application under Platform Settings.",
                             label: "Add as a FIDO trusted app"
                         }
                     },
                     modal: {
                         assertionHint : "I understand and wish to proceed.",
                         header: "Are you sure?",
-                        message: "For validation purposes, Asgardeo requires that the details available under Platform Settings be listed on a public endpoint shared across all Asgardeo organizations.",
+                        message: "For validation purposes, the details available under Platform Settings will be listed on a public endpoint shared across all organizations.",
                         content: ""
                     }
                 },
                 platformSettings: {
                     heading: "Platform Settings",
-                    subTitle: "The following platform specific configurations are needed when enabling client-attestation or trusted app related features",
+                    subTitle: "The following platform specific configurations are needed when enabling client-attestation or trusted app related features.",
                     fields: {
                         android: {
                             heading: "Android",
@@ -1174,11 +1173,11 @@ export const applications: ApplicationsNS = {
                                     }
                                 },
                                 keyHashes: {
-                                    hint: "The SHA256 fingerprints related to the signing certificate of your application",
+                                    hint: "The SHA256 fingerprints related to the signing certificate of your application.",
                                     label: "Key Hashes",
                                     placeholder: "D4:B9:A3",
                                     validations: {
-                                        invalidOrEmpty: "Enter a valid key hash.",
+                                        invalidOrEmpty: "A valid key hash is required for FIDO trusted apps.",
                                         duplicate: "Same key hashes added."
                                     },
                                     tooltip: "Add Thumbprint"
@@ -1764,6 +1763,28 @@ export const applications: ApplicationsNS = {
                         }
                     },
                     heading: "Refresh Token"
+                },
+                subjectToken: {
+                    fields: {
+                        enable: {
+                            hint: "If enabled, this application can be used in the user impersonation flows.",
+                            label: "Enable subject token response type",
+                            validations: {
+                                empty: "This is a required field."
+                            }
+                        },
+                        expiry: {
+                            hint: "Specify the validity period of the <1>subject_token</1> in seconds.",
+                            label: "Subject token expiry time",
+                            placeholder: "Enter the subject token expiry time",
+                            validations: {
+                                empty: "Please fill the subject token expiry time",
+                                invalid: "Subject token expiry time should be in seconds. " +
+                                    "Decimal points and negative numbers are not allowed."
+                            }
+                        }
+                    },
+                    heading: "Subject Token"
                 },
                 requestObjectSignature: {
                     description: "{{productName}} supports receiving an OIDC authentication request as " +