+++ features/admin.core.v1/CHANGELOG.md
@@ -1,5 +1,288 @@
 # @wso2is/admin.core.v1
 
+## 2.22.1
+
+### Patch Changes
+
+- Updated dependencies [[`7f50838d255ffeea0d8c65e1a696f6a923738146`](https://github.com/wso2/identity-apps/commit/7f50838d255ffeea0d8c65e1a696f6a923738146)]:
+  - @wso2is/admin.applications.v1@2.21.29
+  - @wso2is/react-components@2.2.14
+  - @wso2is/i18n@2.5.1
+  - @wso2is/admin.connections.v1@2.20.59
+  - @wso2is/admin.console-settings.v1@2.20.59
+  - @wso2is/admin.extensions.v1@2.21.29
+  - @wso2is/admin.identity-providers.v1@2.20.59
+  - @wso2is/admin.oidc-scopes.v1@2.20.59
+  - @wso2is/admin.organizations.v1@2.20.59
+  - @wso2is/admin.roles.v2@2.20.59
+  - @wso2is/admin.api-resources.v1@2.20.59
+  - @wso2is/admin.api-resources.v2@2.20.59
+  - @wso2is/admin.authorization.v1@2.20.22
+  - @wso2is/admin.branding.v1@2.20.59
+  - @wso2is/admin.certificates.v1@2.20.59
+  - @wso2is/admin.claims.v1@2.20.59
+  - @wso2is/admin.email-and-sms.v1@2.20.59
+  - @wso2is/admin.email-management.v1@2.20.59
+  - @wso2is/admin.email-providers.v1@2.20.59
+  - @wso2is/admin.email-templates.v1@2.20.59
+  - @wso2is/admin.extensions.v2@2.20.59
+  - @wso2is/admin.groups.v1@2.20.59
+  - @wso2is/admin.identity-verification-providers.v1@2.20.59
+  - @wso2is/admin.impersonation.v1@1.1.1
+  - @wso2is/admin.org-insights.v1@2.20.59
+  - @wso2is/admin.organization-discovery.v1@2.20.59
+  - @wso2is/admin.private-key-jwt.v1@2.20.59
+  - @wso2is/admin.remote-repository-configuration.v1@2.20.59
+  - @wso2is/admin.roles.v1@2.20.59
+  - @wso2is/admin.saml2-configuration.v1@2.20.59
+  - @wso2is/admin.secrets.v1@2.20.59
+  - @wso2is/admin.server-configurations.v1@2.21.1
+  - @wso2is/admin.server.v1@2.20.59
+  - @wso2is/admin.session-management.v1@2.20.59
+  - @wso2is/admin.sms-providers.v1@2.20.59
+  - @wso2is/admin.tenants.v1@2.20.59
+  - @wso2is/admin.users.v1@2.20.59
+  - @wso2is/admin.userstores.v1@2.20.59
+  - @wso2is/admin.validation.v1@2.20.59
+  - @wso2is/admin.workflow-approvals.v1@2.20.59
+  - @wso2is/admin.wsfed-configuration.v1@2.20.59
+  - @wso2is/admin.authentication.v1@2.20.59
+
+## 2.22.0
+
+### Minor Changes
+
+- [#6543](https://github.com/wso2/identity-apps/pull/6543) [`a7d0c020b04ba382733280407fc202910f4fc8f4`](https://github.com/wso2/identity-apps/commit/a7d0c020b04ba382733280407fc202910f4fc8f4) Thanks [@Thumimku](https://github.com/Thumimku)! - Add Impersonation support
+
+### Patch Changes
+
+- Updated dependencies [[`a7d0c020b04ba382733280407fc202910f4fc8f4`](https://github.com/wso2/identity-apps/commit/a7d0c020b04ba382733280407fc202910f4fc8f4)]:
+  - @wso2is/admin.server-configurations.v1@2.21.0
+  - @wso2is/admin.impersonation.v1@1.1.0
+  - @wso2is/i18n@2.5.0
+  - @wso2is/admin.applications.v1@2.21.28
+  - @wso2is/admin.branding.v1@2.20.58
+  - @wso2is/admin.claims.v1@2.20.58
+  - @wso2is/admin.console-settings.v1@2.20.58
+  - @wso2is/admin.extensions.v1@2.21.28
+  - @wso2is/admin.identity-providers.v1@2.20.58
+  - @wso2is/admin.private-key-jwt.v1@2.20.58
+  - @wso2is/admin.roles.v1@2.20.58
+  - @wso2is/admin.roles.v2@2.20.58
+  - @wso2is/admin.server.v1@2.20.58
+  - @wso2is/admin.users.v1@2.20.58
+  - @wso2is/admin.validation.v1@2.20.58
+  - @wso2is/admin.api-resources.v1@2.20.58
+  - @wso2is/admin.api-resources.v2@2.20.58
+  - @wso2is/admin.authentication.v1@2.20.58
+  - @wso2is/admin.certificates.v1@2.20.58
+  - @wso2is/admin.connections.v1@2.20.58
+  - @wso2is/admin.email-and-sms.v1@2.20.58
+  - @wso2is/admin.email-management.v1@2.20.58
+  - @wso2is/admin.email-providers.v1@2.20.58
+  - @wso2is/admin.email-templates.v1@2.20.58
+  - @wso2is/admin.extensions.v2@2.20.58
+  - @wso2is/admin.groups.v1@2.20.58
+  - @wso2is/admin.identity-verification-providers.v1@2.20.58
+  - @wso2is/admin.oidc-scopes.v1@2.20.58
+  - @wso2is/admin.org-insights.v1@2.20.58
+  - @wso2is/admin.organization-discovery.v1@2.20.58
+  - @wso2is/admin.organizations.v1@2.20.58
+  - @wso2is/admin.remote-repository-configuration.v1@2.20.58
+  - @wso2is/admin.saml2-configuration.v1@2.20.58
+  - @wso2is/admin.secrets.v1@2.20.58
+  - @wso2is/admin.session-management.v1@2.20.58
+  - @wso2is/admin.sms-providers.v1@2.20.58
+  - @wso2is/admin.tenants.v1@2.20.58
+  - @wso2is/admin.userstores.v1@2.20.58
+  - @wso2is/admin.workflow-approvals.v1@2.20.58
+  - @wso2is/admin.wsfed-configuration.v1@2.20.58
+  - @wso2is/admin.authorization.v1@2.20.21
+
+## 2.21.27
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.authorization.v1@2.20.20
+  - @wso2is/admin.branding.v1@2.20.57
+  - @wso2is/admin.certificates.v1@2.20.57
+  - @wso2is/admin.claims.v1@2.20.57
+  - @wso2is/admin.connections.v1@2.20.57
+  - @wso2is/admin.console-settings.v1@2.20.57
+  - @wso2is/admin.email-and-sms.v1@2.20.57
+  - @wso2is/admin.email-management.v1@2.20.57
+  - @wso2is/admin.email-providers.v1@2.20.57
+  - @wso2is/admin.email-templates.v1@2.20.57
+  - @wso2is/admin.extensions.v1@2.21.27
+  - @wso2is/admin.extensions.v2@2.20.57
+  - @wso2is/admin.groups.v1@2.20.57
+  - @wso2is/admin.identity-providers.v1@2.20.57
+  - @wso2is/admin.identity-verification-providers.v1@2.20.57
+  - @wso2is/admin.oidc-scopes.v1@2.20.57
+  - @wso2is/admin.org-insights.v1@2.20.57
+  - @wso2is/admin.organization-discovery.v1@2.20.57
+  - @wso2is/admin.organizations.v1@2.20.57
+  - @wso2is/admin.private-key-jwt.v1@2.20.57
+  - @wso2is/admin.remote-repository-configuration.v1@2.20.57
+  - @wso2is/admin.roles.v1@2.20.57
+  - @wso2is/admin.roles.v2@2.20.57
+  - @wso2is/admin.saml2-configuration.v1@2.20.57
+  - @wso2is/admin.secrets.v1@2.20.57
+  - @wso2is/admin.server-configurations.v1@2.20.57
+  - @wso2is/admin.server.v1@2.20.57
+  - @wso2is/admin.session-management.v1@2.20.57
+  - @wso2is/admin.sms-providers.v1@2.20.57
+  - @wso2is/admin.tenants.v1@2.20.57
+  - @wso2is/admin.users.v1@2.20.57
+  - @wso2is/admin.userstores.v1@2.20.57
+  - @wso2is/admin.validation.v1@2.20.57
+  - @wso2is/admin.workflow-approvals.v1@2.20.57
+  - @wso2is/admin.wsfed-configuration.v1@2.20.57
+  - @wso2is/react-components@2.2.13
+  - @wso2is/admin.applications.v1@2.21.27
+  - @wso2is/admin.authentication.v1@2.20.57
+  - @wso2is/admin.api-resources.v1@2.20.57
+  - @wso2is/admin.api-resources.v2@2.20.57
+
+## 2.21.26
+
+### Patch Changes
+
+- Updated dependencies [[`a0886211dba54c1c0266c24c5626f61657e6c209`](https://github.com/wso2/identity-apps/commit/a0886211dba54c1c0266c24c5626f61657e6c209)]:
+  - @wso2is/admin.api-resources.v2@2.20.56
+  - @wso2is/admin.applications.v1@2.21.26
+  - @wso2is/admin.roles.v2@2.20.56
+  - @wso2is/admin.connections.v1@2.20.56
+  - @wso2is/admin.console-settings.v1@2.20.56
+  - @wso2is/admin.extensions.v1@2.21.26
+  - @wso2is/admin.identity-providers.v1@2.20.56
+  - @wso2is/admin.oidc-scopes.v1@2.20.56
+  - @wso2is/admin.organizations.v1@2.20.56
+  - @wso2is/admin.api-resources.v1@2.20.56
+  - @wso2is/admin.authentication.v1@2.20.56
+  - @wso2is/admin.branding.v1@2.20.56
+  - @wso2is/admin.certificates.v1@2.20.56
+  - @wso2is/admin.claims.v1@2.20.56
+  - @wso2is/admin.email-and-sms.v1@2.20.56
+  - @wso2is/admin.email-management.v1@2.20.56
+  - @wso2is/admin.email-providers.v1@2.20.56
+  - @wso2is/admin.email-templates.v1@2.20.56
+  - @wso2is/admin.extensions.v2@2.20.56
+  - @wso2is/admin.groups.v1@2.20.56
+  - @wso2is/admin.identity-verification-providers.v1@2.20.56
+  - @wso2is/admin.org-insights.v1@2.20.56
+  - @wso2is/admin.organization-discovery.v1@2.20.56
+  - @wso2is/admin.private-key-jwt.v1@2.20.56
+  - @wso2is/admin.remote-repository-configuration.v1@2.20.56
+  - @wso2is/admin.roles.v1@2.20.56
+  - @wso2is/admin.saml2-configuration.v1@2.20.56
+  - @wso2is/admin.secrets.v1@2.20.56
+  - @wso2is/admin.server-configurations.v1@2.20.56
+  - @wso2is/admin.server.v1@2.20.56
+  - @wso2is/admin.session-management.v1@2.20.56
+  - @wso2is/admin.sms-providers.v1@2.20.56
+  - @wso2is/admin.tenants.v1@2.20.56
+  - @wso2is/admin.users.v1@2.20.56
+  - @wso2is/admin.userstores.v1@2.20.56
+  - @wso2is/admin.validation.v1@2.20.56
+  - @wso2is/admin.workflow-approvals.v1@2.20.56
+  - @wso2is/admin.wsfed-configuration.v1@2.20.56
+
+## 2.21.25
+
+### Patch Changes
+
+- [#6570](https://github.com/wso2/identity-apps/pull/6570) [`82bbbf2303853076342ef80ef7d987bed16df01f`](https://github.com/wso2/identity-apps/commit/82bbbf2303853076342ef80ef7d987bed16df01f) Thanks [@ImalshaG](https://github.com/ImalshaG)! - Load trusted app consent required config and handle consent UI according to it
+
+- Updated dependencies [[`82bbbf2303853076342ef80ef7d987bed16df01f`](https://github.com/wso2/identity-apps/commit/82bbbf2303853076342ef80ef7d987bed16df01f)]:
+  - @wso2is/admin.applications.v1@2.21.25
+  - @wso2is/i18n@2.4.8
+  - @wso2is/admin.connections.v1@2.20.55
+  - @wso2is/admin.console-settings.v1@2.20.55
+  - @wso2is/admin.extensions.v1@2.21.25
+  - @wso2is/admin.identity-providers.v1@2.20.55
+  - @wso2is/admin.oidc-scopes.v1@2.20.55
+  - @wso2is/admin.organizations.v1@2.20.55
+  - @wso2is/admin.roles.v2@2.20.55
+  - @wso2is/admin.api-resources.v1@2.20.55
+  - @wso2is/admin.api-resources.v2@2.20.55
+  - @wso2is/admin.authentication.v1@2.20.55
+  - @wso2is/admin.branding.v1@2.20.55
+  - @wso2is/admin.certificates.v1@2.20.55
+  - @wso2is/admin.claims.v1@2.20.55
+  - @wso2is/admin.email-and-sms.v1@2.20.55
+  - @wso2is/admin.email-management.v1@2.20.55
+  - @wso2is/admin.email-providers.v1@2.20.55
+  - @wso2is/admin.email-templates.v1@2.20.55
+  - @wso2is/admin.extensions.v2@2.20.55
+  - @wso2is/admin.groups.v1@2.20.55
+  - @wso2is/admin.identity-verification-providers.v1@2.20.55
+  - @wso2is/admin.org-insights.v1@2.20.55
+  - @wso2is/admin.organization-discovery.v1@2.20.55
+  - @wso2is/admin.private-key-jwt.v1@2.20.55
+  - @wso2is/admin.remote-repository-configuration.v1@2.20.55
+  - @wso2is/admin.roles.v1@2.20.55
+  - @wso2is/admin.saml2-configuration.v1@2.20.55
+  - @wso2is/admin.secrets.v1@2.20.55
+  - @wso2is/admin.server-configurations.v1@2.20.55
+  - @wso2is/admin.server.v1@2.20.55
+  - @wso2is/admin.session-management.v1@2.20.55
+  - @wso2is/admin.sms-providers.v1@2.20.55
+  - @wso2is/admin.tenants.v1@2.20.55
+  - @wso2is/admin.users.v1@2.20.55
+  - @wso2is/admin.userstores.v1@2.20.55
+  - @wso2is/admin.validation.v1@2.20.55
+  - @wso2is/admin.workflow-approvals.v1@2.20.55
+  - @wso2is/admin.wsfed-configuration.v1@2.20.55
+  - @wso2is/admin.authorization.v1@2.20.19
+
+## 2.21.24
+
+### Patch Changes
+
+- [#6565](https://github.com/wso2/identity-apps/pull/6565) [`1820d54f7dd5fa826c4747964fd2c1af9b7c3d8c`](https://github.com/wso2/identity-apps/commit/1820d54f7dd5fa826c4747964fd2c1af9b7c3d8c) Thanks [@brionmario](https://github.com/brionmario)! - Move routes & views from features to Console source
+
+- Updated dependencies []:
+  - @wso2is/admin.api-resources.v1@2.20.54
+  - @wso2is/admin.api-resources.v2@2.20.54
+  - @wso2is/admin.applications.v1@2.21.24
+  - @wso2is/admin.authentication.v1@2.20.54
+  - @wso2is/admin.branding.v1@2.20.54
+  - @wso2is/admin.certificates.v1@2.20.54
+  - @wso2is/admin.claims.v1@2.20.54
+  - @wso2is/admin.connections.v1@2.20.54
+  - @wso2is/admin.console-settings.v1@2.20.54
+  - @wso2is/admin.email-and-sms.v1@2.20.54
+  - @wso2is/admin.email-management.v1@2.20.54
+  - @wso2is/admin.email-providers.v1@2.20.54
+  - @wso2is/admin.email-templates.v1@2.20.54
+  - @wso2is/admin.extensions.v1@2.21.24
+  - @wso2is/admin.extensions.v2@2.20.54
+  - @wso2is/admin.groups.v1@2.20.54
+  - @wso2is/admin.identity-providers.v1@2.20.54
+  - @wso2is/admin.identity-verification-providers.v1@2.20.54
+  - @wso2is/admin.oidc-scopes.v1@2.20.54
+  - @wso2is/admin.org-insights.v1@2.20.54
+  - @wso2is/admin.organization-discovery.v1@2.20.54
+  - @wso2is/admin.organizations.v1@2.20.54
+  - @wso2is/admin.private-key-jwt.v1@2.20.54
+  - @wso2is/admin.remote-repository-configuration.v1@2.20.54
+  - @wso2is/admin.roles.v1@2.20.54
+  - @wso2is/admin.roles.v2@2.20.54
+  - @wso2is/admin.saml2-configuration.v1@2.20.54
+  - @wso2is/admin.secrets.v1@2.20.54
+  - @wso2is/admin.server-configurations.v1@2.20.54
+  - @wso2is/admin.server.v1@2.20.54
+  - @wso2is/admin.session-management.v1@2.20.54
+  - @wso2is/admin.sms-providers.v1@2.20.54
+  - @wso2is/admin.tenants.v1@2.20.54
+  - @wso2is/admin.users.v1@2.20.54
+  - @wso2is/admin.userstores.v1@2.20.54
+  - @wso2is/admin.validation.v1@2.20.54
+  - @wso2is/admin.workflow-approvals.v1@2.20.54
+  - @wso2is/admin.wsfed-configuration.v1@2.20.54
+
 ## 2.21.23
 
 ### Patch Changes