+++ modules/form/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @wso2is/form
 
+## 2.0.73
+
+### Patch Changes
+
+- Updated dependencies [[`7f50838d255ffeea0d8c65e1a696f6a923738146`](https://github.com/wso2/identity-apps/commit/7f50838d255ffeea0d8c65e1a696f6a923738146)]:
+  - @wso2is/react-components@2.2.14
+
+## 2.0.72
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/react-components@2.2.13
+
 ## 2.0.71
 
 ### Patch Changes